# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E2BR3 Backend API",
    "description": "OpenAPI documentation for the Axum-based E2BR3 backend. Authenticated API routes use the HTTP-only `auth-token` cookie. Internal machine-to-machine routes use the `x-callback-token` header.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/presave-templates": {
      "get": {
        "tags": [
          "presave-templates"
        ],
        "operationId": "list_presave_templates",
        "parameters": [
          {
            "name": "activeOnly",
            "in": "query",
            "description": "Filter active templates only",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Presave templates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericDataResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      },
      "post": {
        "tags": [
          "presave-templates"
        ],
        "operationId": "create_presave_template",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Presave template created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericDataResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      }
    },
    "/api/presave-templates/{id}": {
      "get": {
        "tags": [
          "presave-templates"
        ],
        "operationId": "get_presave_template",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Template ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Presave template",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericDataResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "presave-templates"
        ],
        "operationId": "delete_presave_template",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Template ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Presave template deleted"
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      },
      "patch": {
        "tags": [
          "presave-templates"
        ],
        "operationId": "update_presave_template",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Template ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Presave template updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericDataResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      }
    },
    "/api/presave-templates/{id}/audit": {
      "get": {
        "tags": [
          "presave-templates"
        ],
        "operationId": "list_presave_template_audits",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Template ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Presave template audit trail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericDataResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenericDataRequest": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object"
          }
        }
      },
      "GenericDataResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object"
          }
        }
      },
...
    },
    "securitySchemes": {
      "auth_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth-token"
      },
      "callback_token": {
        "type": "apiKey",
        "in": "header",
        "name": "x-callback-token"
      }
    }
  },
...
}